Separate suggestion extension from contentlet block extension
